build: generate configure in build dir

+++ autogen.sh
@@ -1,5 +1,18 @@
 #!/bin/sh
 
+REQ_FILES="configure.ac getversion VERSION m4"
+
+if [ -n "$1" ]; then
+    for i in $REQ_FILES; do
+	ln -sf "$1"/$i $i
+    done
+    mkdir -p src/include
+    if [ -f "$1"/configure ]; then
+	echo "cp -fp "$1"/configure configure"
+	cp -fp "$1"/configure configure
+	exit $?
+    fi
+fi
 echo "Generating toplevel configure script..."
 rm -f aclocal.m4
 if ! autoreconf -I m4 --install --force ; then
